ALte Klassen gelöscht und die Hauptklasse in "Main" umbenannt
